requirements.txt
autopep8==1.4.4
certifi==2019.6.16
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
olefile==0.46
Pillow==6.1.0
pycodestyle==2.5.0
six==1.12.0
wincertstore==0.2
